Two implementations of reusable todo design
1. Reactive extensions (RxSwift) version. App state is a Variable which can be
observed, with any number of subscribers watching for changes to the value of the Variable.
These subscriptions are used to update the view and could be used to watch for state changes
within other parts of the app as well.
2. Redux version. App state is contained within a store and can only be modified via the reducer
functions owned by the store. If we want to modify the list, we dispatch an action which in
turn updates the state, and UI is subscribing to these state changes to update its view.